python-workers/app.py
# Python Workers для обработки видео
# TODO: Установите зависимости командой: pip install -r requirements.txt

"""
Этот модуль будет использоваться для:
- Скачивания видео через yt-dlp
- Распознавания речи через Whisper
- Генерации голоса через gTTS
- Обработки видео через MoviePy
- Перевода через deep-translator

Пока что это заглушка. Функционал будет добавлен позже.
"""

try:
    from flask import Flask, request, jsonify
    from flask_cors import CORS
    import os
    from dotenv import load_dotenv

    load_dotenv()

    app = Flask(__name__)
    CORS(app)

    @app.route('/health', methods=['GET'])
    def health():
        return jsonify({
            'status': 'OK',
            'service': 'Python Video Worker'
        })

    @app.route('/generate', methods=['POST'])
    def generate_video():
        """
        Генерация видео с переводом
        TODO: Реализовать интеграцию с Whisper, MoviePy, gTTS
        """
        data = request.json
        video_id = data.get('videoId')
        target_languages = data.get('targetLanguages', [])
        
        return jsonify({
            'success': True,
            'message': 'В разработке',
            'taskId': 'task_' + video_id,
            'videoId': video_id,
            'languages': target_languages
        })

    if __name__ == '__main__':
        port = int(os.getenv('PYTHON_WORKER_PORT', 5000))
        app.run(host='0.0.0.0', port=port, debug=True)
        
except ImportError as e:
    print("⚠️  Python зависимости не установлены")
    print("📦 Установите их командой: pip install -r requirements.txt")
    print(f"Ошибка: {e}")
    print("\nПока что Python workers не требуются для работы приложения.")
    print("Основной функционал работает на Node.js + React")

