python-workers/requirements.txt
google-api-python-client
yt-dlp
moviepy
whisper
openai-whisper
gtts
deep-translator
flask
flask-cors
python-dotenv
redis
